fix: set int to int64
Change int to int64 in structures to match db types

+++ api/types/message.go
@@ -3,9 +3,9 @@ package types
 import "time"
 
 type Message struct {
-	ID        int
+	ID        int64
 	Content   string    `json:"content"`
-	UserID    int       `json:"user_id"`
+	UserID    int64     `json:"user_id"`
 	User      User      `json:"user"`
 	CreatedAt time.Time `json:"created_at"`
 }
